requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.10.4
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.0
idna==3.10
Jinja2==3.1.4
jiter==0.6.1
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.1
numpy==2.1.2
openai==0.28.0
packaging==24.1
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-telegram-bot==21.6
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
tokenizers==0.20.1
torch==2.5.0
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
yarl==1.15.5
